+++ src/mito2/src/sst/parquet/reader.rs
@@ -34,13 +34,12 @@ use datatypes::arrow::record_batch::RecordBatch;
 use datatypes::data_type::ConcreteDataType;
 use datatypes::prelude::DataType;
 use futures::StreamExt;
-use mito_codec::row_converter::{PrimaryKeyCodec, build_primary_key_codec};
+use mito_codec::row_converter::build_primary_key_codec;
 use object_store::ObjectStore;
 use parquet::arrow::ProjectionMask;
 use parquet::arrow::arrow_reader::{ArrowReaderMetadata, ArrowReaderOptions, RowSelection};
 use parquet::arrow::async_reader::{ParquetRecordBatchStream, ParquetRecordBatchStreamBuilder};
 use parquet::file::metadata::{PageIndexPolicy, ParquetMetaData};
-use parquet::schema::types::SchemaDescriptor;
 use partition::expr::PartitionExpr;
 use snafu::ResultExt;
 use store_api::codec::PrimaryKeyEncoding;
@@ -82,7 +81,7 @@ use crate::sst::parquet::flat_format::FlatReadFormat;
 use crate::sst::parquet::format::need_override_sequence;
 use crate::sst::parquet::metadata::MetadataLoader;
 use crate::sst::parquet::prefilter::{
-    PrefilterContextBuilder, execute_prefilter, is_usable_primary_key_filter,
+    PrefilterContextBuilder, build_reader_filter_plan, execute_prefilter,
 };
 use crate::sst::parquet::read_columns::{
     ParquetReadColumns, ProjectionMaskPlan, build_projection_plan,
@@ -96,11 +95,6 @@ const INDEX_TYPE_FULLTEXT: &str = "fulltext";
 const INDEX_TYPE_INVERTED: &str = "inverted";
 const INDEX_TYPE_BLOOM: &str = "bloom filter";
 const INDEX_TYPE_VECTOR: &str = "vector";
-
-struct ReaderFilterPlan {
-    remaining_simple_filters: Vec<SimpleFilterContext>,
-    prefilter_builder: Option<PrefilterContextBuilder>,
-}
 
 macro_rules! handle_index_error {
     ($err:expr, $file_handle:expr, $index_type:expr) => {
@@ -466,8 +460,10 @@ impl ParquetReaderBuilder {
 
         let codec = build_primary_key_codec(read_format.metadata());
 
-        let filter_plan = self.build_filter_plan(
-            &region_meta,
+        let filter_plan = build_reader_filter_plan(
+            self.predicate.as_ref(),
+            self.expected_metadata.as_deref(),
+            self.pre_filter_mode,
             &read_format,
             parquet_meta.file_metadata().schema_descr(),
             &codec,
@@ -508,112 +504,6 @@ impl ParquetReaderBuilder {
         metrics.build_cost += start.elapsed();
 
         Ok(Some((context, selection)))
-    }
-
-    fn build_filter_plan(
-        &self,
-        region_meta: &RegionMetadataRef,
-        read_format: &ReadFormat,
-        parquet_schema: &SchemaDescriptor,
-        codec: &Arc<dyn PrimaryKeyCodec>,
-    ) -> ReaderFilterPlan {
-        let Some(predicate) = &self.predicate else {
-            return ReaderFilterPlan {
-                remaining_simple_filters: Vec::new(),
-                prefilter_builder: None,
-            };
-        };
-
-        let mut simple_filters = Vec::new();
-        let mut prefilter_simple_filters = Vec::new();
-        let mut remaining_simple_filters = Vec::new();
-        let mut prefilter_physical_filters = Vec::new();
-        let mut primary_key_filters = Vec::new();
-
-        let field_prefilter_enabled = self.pre_filter_mode == PreFilterMode::All;
-        let supports_pk_prefilter = read_format
-            .as_flat()
-            .is_some_and(|flat_format| !flat_format.batch_has_raw_pk_columns());
-
-        for expr in predicate.exprs() {
-            if let Some(filter) =
-                SimpleFilterContext::new_opt(region_meta, self.expected_metadata.as_deref(), expr)
-            {
-                simple_filters.push(filter);
-                continue;
-            }
-
-            if let Some(filter) = PhysicalFilterContext::new_opt(
-                region_meta,
-                self.expected_metadata.as_deref(),
-                read_format,
-                expr,
-            ) {
-                let usable_prefilter = (field_prefilter_enabled
-                    || filter.semantic_type() != SemanticType::Field)
-                    && read_format
-                        .arrow_schema()
-                        .column_with_name(filter.column_name())
-                        .is_some();
-                if usable_prefilter {
-                    prefilter_physical_filters.push(filter);
-                }
-            }
-        }
-
-        for filter_ctx in simple_filters.iter().cloned() {
-            let Some(filter) = filter_ctx.filter().as_filter() else {
-                remaining_simple_filters.push(filter_ctx);
-                continue;
-            };
-
-            let direct_prefilter = (field_prefilter_enabled
-                || filter_ctx.semantic_type() != SemanticType::Field)
-                && read_format
-                    .arrow_schema()
-                    .column_with_name(filter.column_name())
-                    .is_some();
-            if direct_prefilter {
-                prefilter_simple_filters.push(filter_ctx);
-                continue;
-            }
-
-            let pk_prefilter = supports_pk_prefilter
-                && is_usable_primary_key_filter(
-                    region_meta,
-                    self.expected_metadata.as_deref(),
-                    filter,
-                );
-            if pk_prefilter {
-                primary_key_filters.push(filter.clone());
-                prefilter_simple_filters.push(filter_ctx);
-            } else {
-                remaining_simple_filters.push(filter_ctx);
-            }
-        }
-
-        let primary_key_filters =
-            (!primary_key_filters.is_empty()).then_some(Arc::new(primary_key_filters));
-        let prefilter_builder = PrefilterContextBuilder::new(
-            read_format,
-            codec,
-            primary_key_filters,
-            prefilter_simple_filters,
-            prefilter_physical_filters,
-            parquet_schema,
-        );
-
-        if prefilter_builder.is_some() {
-            ReaderFilterPlan {
-                remaining_simple_filters,
-                prefilter_builder,
-            }
-        } else {
-            ReaderFilterPlan {
-                remaining_simple_filters: simple_filters,
-                prefilter_builder: None,
-            }
-        }
     }
 
     fn is_same_region_partition(
@@ -2426,85 +2316,6 @@ mod tests {
             .primary_key(vec![10, 1]);
 
         Arc::new(builder.build().unwrap())
-    }
-
-    fn new_test_parquet_reader_builder(
-        predicate: Predicate,
-        pre_filter_mode: PreFilterMode,
-    ) -> ParquetReaderBuilder {
-        let object_store = ObjectStore::new(Memory::default()).unwrap().finish();
-        let file_handle = sst_file_handle(0, 1);
-
-        ParquetReaderBuilder::new(
-            "test_table".to_string(),
-            PathType::Bare,
-            file_handle,
-            object_store,
-        )
-        .predicate(Some(predicate))
-        .pre_filter_mode(pre_filter_mode)
-    }
-
-    #[test]
-    fn test_build_filter_plan_keeps_field_filters_for_prune_reader_in_skip_fields_mode() {
-        let metadata: RegionMetadataRef = Arc::new(sst_region_metadata());
-        let read_format = ReadFormat::new_flat(
-            metadata.clone(),
-            metadata.column_metadatas.iter().map(|c| c.column_id),
-            None,
-            "test",
-            true,
-        )
-        .unwrap();
-        let parquet_schema = parquet::arrow::ArrowSchemaConverter::new()
-            .convert(read_format.arrow_schema())
-            .unwrap();
-        let codec = build_primary_key_codec(metadata.as_ref());
-        let builder = new_test_parquet_reader_builder(
-            Predicate::new(vec![
-                col("tag_0").eq(lit("a")),
-                col("field_0").gt(lit(1_u64)),
-            ]),
-            PreFilterMode::SkipFields,
-        );
-
-        let plan = builder.build_filter_plan(&metadata, &read_format, &parquet_schema, &codec);
-
-        assert!(plan.prefilter_builder.is_some());
-        assert_eq!(plan.remaining_simple_filters.len(), 1);
-        let remaining_filter = plan.remaining_simple_filters[0]
-            .filter()
-            .as_filter()
-            .unwrap();
-        assert_eq!(remaining_filter.column_name(), "field_0");
-    }
-
-    #[test]
-    fn test_build_filter_plan_uses_pk_prefilter_for_unprojected_tag_filters() {
-        let metadata: RegionMetadataRef = Arc::new(sst_region_metadata());
-        let field_0 = metadata.column_by_name("field_0").unwrap().column_id;
-        let ts = metadata.time_index_column().column_id;
-        let read_format = ReadFormat::new_flat(
-            metadata.clone(),
-            [field_0, ts].into_iter(),
-            None,
-            "test",
-            true,
-        )
-        .unwrap();
-        let parquet_schema = parquet::arrow::ArrowSchemaConverter::new()
-            .convert(read_format.arrow_schema())
-            .unwrap();
-        let codec = build_primary_key_codec(metadata.as_ref());
-        let builder = new_test_parquet_reader_builder(
-            Predicate::new(vec![col("tag_0").eq(lit("a"))]),
-            PreFilterMode::All,
-        );
-
-        let plan = builder.build_filter_plan(&metadata, &read_format, &parquet_schema, &codec);
-
-        assert!(plan.prefilter_builder.is_some());
-        assert!(plan.remaining_simple_filters.is_empty());
     }
 
     #[test]
